moved preview to a modal
